// server/src/domain/enums/routes.enum.ts
export enum CompanyRoutes {
  PROFILE = '/profile',
  PROFILE_BY_ID = '/profile/:profileId',
  REAPPLY_VERIFICATION = '/reapply-verification',
  UPLOAD_LOGO = '/upload/logo',
  UPLOAD_BUSINESS_LICENSE = '/upload/business-license',
  DELETE_UPLOAD = '/upload/delete',
  SUBSCRIPTION_PLANS = '/subscription-plans',
  DASHBOARD = '/dashboard',
  CONTACT = '/contact',
  TECH_STACKS = '/tech-stacks',
  TECH_STACKS_ID = '/tech-stacks/:id',
  OFFICE_LOCATIONS = '/office-locations',
  OFFICE_LOCATIONS_ID = '/office-locations/:id',
  BENEFITS = '/benefits',
  BENEFITS_ID = '/benefits/:id',
  WORKPLACE_PICTURES = '/workplace-pictures',
  WORKPLACE_PICTURES_ID = '/workplace-pictures/:id',
  WORKPLACE_PICTURES_UPLOAD = '/workplace-pictures/upload',
  SUBSCRIPTIONS_ACTIVE = '/subscriptions/active',
  SUBSCRIPTIONS_PAYMENT_HISTORY = '/subscriptions/payment-history',
  SUBSCRIPTIONS_CREATE_CHECKOUT = '/subscriptions/create-checkout-session',
  SUBSCRIPTIONS_CANCEL = '/subscriptions/cancel',
  SUBSCRIPTIONS_RESUME = '/subscriptions/resume',
  SUBSCRIPTIONS_CHANGE_PLAN = '/subscriptions/change-plan',
  SUBSCRIPTIONS_BILLING_PORTAL = '/subscriptions/billing-portal',
  JOBS = '/jobs',
  JOBS_ID = '/jobs/:id',
  JOBS_ID_STATUS = '/jobs/:id/status',
  APPLICATIONS = '/applications',
  APPLICATIONS_BULK_UPDATE = '/applications/bulk-update',
  APPLICATIONS_ID = '/applications/:id',
  APPLICATIONS_ID_STAGE = '/applications/:id/stage',
  APPLICATIONS_ID_SCORE = '/applications/:id/score',
  APPLICATIONS_ID_INTERVIEWS = '/applications/:id/interviews',
  APPLICATIONS_ID_INTERVIEWS_INTERVIEW_ID = '/applications/:id/interviews/:interviewId',
  APPLICATIONS_ID_INTERVIEWS_INTERVIEW_DELETE = '/applications/:id/interviews/:interviewId',
  APPLICATIONS_ID_INTERVIEWS_FEEDBACK = '/applications/:id/interviews/:interviewId/feedback',
}